+++ app/views/question/categorizations/_categorization.xml.erb
@@ -1,9 +1,9 @@
-<item ident="<%= matching.item_ident %>" title="Question <%= matching.id %>" >
+<item ident="<%= categorization.item_ident %>" title="Question <%= categorization.id %>" >
   <itemmetadata>
     <qtimetadata>
       <qtimetadatafield>
         <fieldlabel>question_type</fieldlabel>
-        <fieldentry>matching_question</fieldentry>
+        <fieldentry>categorization_question</fieldentry>
       </qtimetadatafield>
       <qtimetadatafield>
         <fieldlabel>points_possible</fieldlabel>
@@ -11,48 +11,48 @@
       </qtimetadatafield>
       <qtimetadatafield>
         <fieldlabel>original_answer_ids</fieldlabel>
-        <fieldentry><%= matching.qti_choices.map(&:ident).join(',') %></fieldentry>
+        <fieldentry></fieldentry>
       </qtimetadatafield>
       <qtimetadatafield>
         <fieldlabel>assessment_question_identifierref</fieldlabel>
-        <fieldentry><%= matching.assessment_question_identifierref %></fieldentry>
+        <fieldentry><%= categorization.assessment_question_identifierref %></fieldentry>
       </qtimetadatafield>
     </qtimetadata>
   </itemmetadata>
   <presentation>
     <material>
-      <mattext texttype="text/plain"><%= matching.text %></mattext>
+      <mattext texttype="text/plain"><%= categorization.text %></mattext>
     </material>
-    <% matching.qti_choices.each do |choice| %>
-      <response_lid ident="<%= choice.ident %>">
+    <%- categorization.qti_choices.each do |choice| %>
+      <response_lid ident="<%= choice.ident %>" rcardinality="Multiple">
 	<material>
           <mattext texttype="text/plain"><%= choice.text %></mattext>
 	</material>
 	<render_choice>
-	  <% matching.qti_responses.each do |qresponse| %>
+	  <%- categorization.qti_responses.each do |qresponse| %>
             <response_label ident="<%= qresponse.ident %>">
               <material>
 		<mattext><%= qresponse.text %></mattext>
               </material>
             </response_label>
-	  <% end %>
+	  <%- end %>
       </render_choice>
       </response_lid>
-    <% end %>
+    <%- end %>
   </presentation>
   <resprocessing>
     <outcomes>
-      <decvar maxvalue="<%= matching.qti_max_value %>" minvalue="0" varname="SCORE" vartype="Decimal"/>
+      <decvar maxvalue="<%= categorization.qti_max_value %>" minvalue="0" varname="SCORE" vartype="Decimal"/>
     </outcomes>
-    <% matching.qti_response_conditions.each do |condition| %>
+    <%- categorization.qti_response_conditions.each do |condition| %>
       <respcondition>
 	<conditionvar>
 	  <%- condition.responses.each do |response| %>
-            <varequal respident="<%= condition.choice_ident %>"><%= response.ident %></varequal>
+          <varequal respident="<%= condition.choice_ident %>"><%= response.ident %></varequal>
 	  <%- end %>
 	</conditionvar>
 	<setvar varname="SCORE" action="Add"><%= condition.value %></setvar>
       </respcondition>
-    <% end %>
+    <%- end %>
   </resprocessing>
 </item>
