+++ spec/models/question/matching_spec.rb
@@ -8,24 +8,4 @@ RSpec.describe Question::Matching do
   its(:type_label) { is_expected.to eq("Question") }
   its(:type_name) { is_expected.to eq("Matching") }
   its(:qti_max_value) { is_expected.to be_a(Integer) }
-
-  let(:instance) { FactoryBot.build(:question_matching) }
-
-  describe '#qti_choices' do
-    it "is an Array of Choice objects" do
-      expect(instance.qti_choices.all? { |r| r.is_a?(described_class::Choice) }).to be_truthy
-    end
-  end
-
-  describe '#qti_response_conditions' do
-    it "is an Array of ResponseCondition objects" do
-      expect(instance.qti_response_conditions.all? { |r| r.is_a?(described_class::ResponseCondition) }).to be_truthy
-    end
-  end
-
-  describe '#qti_responses' do
-    it "is an Array of Response objects" do
-      expect(instance.qti_responses.all? { |r| r.is_a?(described_class::Response) }).to be_truthy
-    end
-  end
 end
