+++ spec/shared_examples.rb
@@ -11,6 +11,7 @@ RSpec.shared_examples 'a Question' do |valid: true, export_as_xml: false, test_t
   its(:has_parts?) { is_expected.to eq(has_parts) }
   its(:question) { is_expected.to be_a(described_class) }
   its(:question) { is_expected.to eq(subject) }
+  its(:qti_max_value) { is_expected.to eq(100) }
 
   if test_type_name_to_class
     describe '.type_name_to_class' do
@@ -158,6 +159,28 @@ RSpec.shared_examples 'a Matching Question' do
         else
           it { is_expected.not_to be_valid }
         end
+      end
+    end
+  end
+
+  describe 'QTI Exporting' do
+    let(:instance) { FactoryBot.build(:question_matching) }
+
+    describe '#qti_choices' do
+      it "is an Array of Choice objects" do
+        expect(instance.qti_choices.all? { |r| r.is_a?(described_class::Choice) }).to be_truthy
+      end
+    end
+
+    describe '#qti_response_conditions' do
+      it "is an Array of ResponseCondition objects" do
+        expect(instance.qti_response_conditions.all? { |r| r.is_a?(described_class::ResponseCondition) }).to be_truthy
+      end
+    end
+
+    describe '#qti_responses' do
+      it "is an Array of Response objects" do
+        expect(instance.qti_responses.all? { |r| r.is_a?(described_class::Response) }).to be_truthy
       end
     end
   end
